requirements.txt
amqp==2.4.0
celery==4.2.1
Django==2.1.5
django-gm2m==0.6.1
redis==3.0.1
